Extract truncate0x utility fn
