+++ src/utils/generic/index.ts
@@ -2,6 +2,7 @@ import checkIfContainsAll from "./checkIfContainsAll";
 import findUnlockedCids from "./findUnlockedCids";
 import getCurrentYear from "./getCurrentYear";
 import findUnlockableByUuid from "./findUnlockableByUuid";
+import truncate0x from "./truncate0x";
 import cn from "./cn";
 
 export {
@@ -9,5 +10,6 @@ export {
   findUnlockedCids,
   findUnlockableByUuid,
   getCurrentYear,
+  truncate0x,
   cn,
 };
